sky-control/requirements.txt
openenv-core
transformers
torch
openai
